Create separate CalendarView and SelectTaskModeView
